% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_functions.R
\name{apply_acs_tm}
\alias{apply_acs_tm}
\title{Apply ACS tranformation matrix to rigid body transformation}
\usage{
apply_acs_tm(rbt, acs_tm)
}
\arguments{
\item{rbt}{2- or 3-D matrix where D1, D2 are 4x4 transformation matrixes and
D3 is frames}

\item{acs_tm}{4x4 transformation matrix of ACS relative to bone model}
}
\value{
new transformation matrix, same dimensions as input RBT
}
\description{
The rigid body transformations that are output from XMALab describe a
transformation from CT space to Cube space. \code{apply_acs_tm} applies the
ACS transformation matrix to the RBTs, such tht the new_tm describes the
transformation of a bone from anatomical (ACS) space to Cube space, and thus
can be used to calculate the relative motion of points and other bones. In
short, this function multiplies the RBT by the ACS TM for every frame.
}
\examples{
\dontrun{
new_rbt <- apply_acs_tm(rbts[,,,1], acs_tms[,,1])
}
}
\author{
Kara L. Feilich, J.D. Laurence-Chasen
}
