% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_functions.R
\name{calculate_relative_motion}
\alias{calculate_relative_motion}
\title{Calculate relative motion of two bodies}
\usage{
calculate_relative_motion(reference_element_xform, mobile_element_xform)
}
\arguments{
\item{reference_element_xform}{an array of reference/proximal rigid body
transformation (transformation matrix)}

\item{mobile_element_xform}{an array of either 3D point/s position or mobile
element's rigid body transformation (transformation matrix). In the case of
3D points, format is D1 = points, D2 = [x, y, z] positions, and D3 is frames;
in the case of a transformation matrix, D1 and D2 are 4x4 TMs, and D3 is frames.}
}
\value{
array of relative motion of same dimensions as mobile_element_xform
}
\description{
This function, similar to oRel in Maya, calculates the relative morion of
either 'mobile' 3D point/s or rigid body relative to 'proximal'/reference
rigid body. The function is written to handle multiple frames/timepoints of
data (see arguments for format), and will automatically detect which frames have
data and pad the output with NaNs accordingly.
}
\details{
The math underlying the calculation ultimately boils down to multiplying
the inverse of the reference transformation matrix by the mobile element's
position (XYZ or transformation matrix).
}
\examples{
\dontrun{

}
}
\author{
Kara L. Feilich, J.D. Laurence-Chasen
}
