% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.R
\name{export_tm}
\alias{export_tm}
\title{Converts the RBT array output by calculate_relative_motion( ) to 2D matrix}
\usage{
export_tm(tm)
}
\arguments{
\item{tm}{a 3D transformation matrix}
}
\value{
a 2D matrix of dimensions nFrames x 16
}
\description{
\code{export_tm} converts the transformation array output by
\code{calculate_relative_motion} into a 2D matrix w/ dimensions nFrames x 16
suitable for export to .csv that can be read into Maya.
}
\examples{
\dontrun{
 }
}
\author{
Kara Feilich
}
