% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.R
\name{export_xyz_points}
\alias{export_xyz_points}
\title{Converts the points array output by calculate_relative_motion( ) to 2D matrix}
\usage{
export_xyz_points(xyz_points)
}
\arguments{
\item{xyz_points}{a 3D array}
}
\value{
a 2D matrix of size nFrames x (3 x nLandmarks)
}
\description{
\code{export_xyz_points} converts the transformation array output by
\code{calculate_relative_motion} into a 2D matrix w/ dimensions nFrames x (3 x
nLandmarks) suitable for export to .csv that can be read into Maya.
}
\examples{
\dontrun{
 }
}
\author{
Kara Feilich
}
