% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{import_acs_rel_bone}
\alias{import_acs_rel_bone}
\title{Converts the ACS transformation matrix file from XMALab into a 4 x 4 matrix}
\usage{
import_acs_rel_bone(filename)
}
\arguments{
\item{filename}{a character string specifying the RBT filepath from Maya
following tutorial instructions)}
}
\value{
a 2-D (4x4) transformation matrix that describes the position of the
ACS relative to a bone model
}
\description{
\code{import_acs_rel_bone} takes the ACS transformation matrix output by
XMALab, and converts it into a matrix of dimensions [4 x 4].
}
\examples{
\dontrun{
import_acs_rel_bone <- import_acs_rel_bone("Craniums_acs_tm.csv")
}
}
\author{
Kara L. Feilich, J.D. Laurence-Chasen
}
