% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{import_rbt}
\alias{import_rbt}
\title{Takes the RBT file from XMALab and converts it into a 4-D array}
\usage{
import_rbt(filename)
}
\arguments{
\item{filename}{a character string specifying the filepath}
}
\value{
a 4-D array where the first two dimensions are the rows and
columns of a 4x4 transformation matrix, D3 is frames, and D4
is rigid bodies.
}
\description{
\code{import_rbt} takes the RBT file output by XMALab, and converts it into an
array of dimensions [4 x 4 x n_frames x n_rigid_bodies]. The 4x4
component is the transformation matrix. If the RBT file contains only one
RBT (as when "Single Files" is checked during XMALab export,
\code{dim(rbt_array)[4] == 1}.
}
\examples{
\dontrun{
rbt_array <- import_rbt("Trial01_RBT.csv")
}
}
\author{
Kara L. Feilich, J.D. Laurence-Chasen
}
