% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{import_xyz_points}
\alias{import_xyz_points}
\title{Import xyz (3D) points}
\usage{
import_xyz_points(filename)
}
\arguments{
\item{filename}{a character string specifying the 3D points filepath from XMALab}
}
\value{
a 3-dimensional array where D1 is points (as in framespec file), D2
is [X,Y,Z] positions, and D3 is frames
}
\description{
\code{import_xyz_points} takes the 3D points file output from XMALab and
converts it into a 3-D array of dimensions [n_points x 3 x n_frames], where
columns are [x, y, z] positions.
}
\examples{
\dontrun{
xyz_array <- import_xyz_points("Trial01_3DPoints.csv")}
}
\author{
Kara L. Feilich, J.D. Laurence-Chasen
}
